almost done

+++ service.py
@@ -11,18 +11,20 @@ def add_pokemon(id, name, height, weight, types):
                 cursor.execute("insert into type(name, pokemon_id) values('{}', {})".format(types[i],id))
             connection.commit()
     except:
-        print("DB Error")
+        return "DB Error"
 
 
 
 def delete_pokemon(pokemon_name):
     try:
         with connection.cursor() as cursor:
+            if cursor.execute("select * from pokemon where name=(%s)",(pokemon_name))==0:
+                return "pokemon doesn't exist"
             cursor.execute("delete from type where pokemon_id = (select id from pokemon where name = '{}')".format(pokemon_name))
             cursor.execute("delete from pokemon where name = '{}'".format(pokemon_name))
             connection.commit()
     except:
-        print("DB Error")
+        return "DB Error"
 
 
 def heaviest_pokemon():
@@ -40,31 +42,51 @@ def find_by_type(type):
         with connection.cursor() as cursor:
             cursor.execute( "select pokemon.name FROM type join pokemon on pokemon.id=type.pokemon_id where type.name=(%s);",(type))
             result = cursor.fetchall()
+            if not result:
+                return "type does not exist"
             return [i['name'] for i in result]
     except:
-        print("Error")
-
+        return "DB Error"
 
 
 def find_owners(pokemon_name):
     try:
         with connection.cursor() as cursor:
+            if cursor.execute("select * from pokemon where name=(%s)",(pokemon_name))==0:
+                return "pokemon doesn't exist"
             cursor.execute( "SELECT ownedBy.owner_name FROM pokemon JOIN ownedBy ON pokemon.id = ownedBy.pokemon_id where pokemon.name=(%s);",(pokemon_name))
             result = cursor.fetchall()
             return [i['owner_name'] for i in result]
     except:
-        print("Error")
+        return "DB Error"
 
 
 
 def find_roster(trainer_name):
     try:
         with connection.cursor() as cursor:
+            if cursor.execute("select * from owner where name=(%s)",(trainer_name))==0:
+                return "trainer doesn't exist"
             cursor.execute( "SELECT pokemon.name FROM pokemon JOIN ownedBy ON pokemon.id = ownedBy.pokemon_id where ownedBy.owner_name=(%s);",(trainer_name))
             result = cursor.fetchall()
             return [i['name'] for i in result]
     except:
-        print("Error")
+        return "DB Error"
+
+
+
+def update_types(name,types):
+    try:
+        with connection.cursor() as cursor:
+            cursor.execute( "SELECT id FROM pokemon where name=(%s)",(name))
+            id=cursor.fetchall()[0].get('id')
+            for i in types:
+                cursor.execute("INSERT INTO type values(%s,%s)",(i,id))
+            connection.commit()
+            return True
+    except:
+        return False
+
 
 
 def finds_most_owned():
@@ -84,17 +106,6 @@ def finds_most_owned():
 
 
 
-def update_types(name,types):
-    try:
-        with connection.cursor() as cursor:
-            for i in types:
-                cursor.execute( "SELECT id FROM pokemon where name=(%s)",(name))
-                id=cursor.fetchall()[0].get('id')
-                cursor.execute("INSERT INTO type values(%s,%s)",(i,id))
-            connection.commit()
-            return True
-    except:
-        return False
 
 
 def evolve_pokemon(pokemon_name, trainer, evolves_to):
@@ -105,19 +116,22 @@ def evolve_pokemon(pokemon_name, trainer, evolves_to):
             
             cursor.execute( "SELECT id FROM pokemon where name=(%s)",(pokemon_name))
             pokemon_id = cursor.fetchall()[0]['id']
+
+            if cursor.execute( "SELECT * FROM owner where name=(%s)",(trainer))==0:
+                return "trainer doesn't exist"
+
             cursor.execute("select * from ownedby where pokemon_id = (%s) and owner_name = (%s)", (pokemon_id, trainer ))
             does_exist = cursor.fetchall()
             if not does_exist:
-                return False
+                return "the owner does not have this pokemon"
             cursor.execute("select * from ownedby where pokemon_id = (%s) and owner_name = (%s)", (evolves_id, trainer ))
             has_evolve = cursor.fetchall()
             if has_evolve:
-                return True
+                return "evolved pokemon already exists"
             cursor.execute("UPDATE ownedBy SET pokemon_id =(%s)  where pokemon_id = (%s) and owner_name = (%s)", (evolves_id, pokemon_id, trainer ))
             connection.commit()
-            return True
     except:
-        return False
+        return "DB Error"
 
 
 def does_pokemon_exist(pokemon_name):
